backend/requirements.txt
Flask
flask-cors==4.0.0
python-dotenv==1.0.0
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
pip>=21.0
google-auth-oauthlib==1.2.0
pymongo==4.5.0
requests==2.31.0
gunicorn==21.2.0
flask-socketio==5.3.6
eventlet==0.33.3
uuid==1.30
PyJWT